+++ lib/shrine/plugins/multi_delete.rb
@@ -5,9 +5,9 @@ class Shrine
     #
     #     plugin :multi_delete
     #
-    # This plugin allows you pass an array of files to `Shrine.delete`.
+    # This plugin allows you pass an array of files to `Shrine#delete`.
     #
-    #     Shrine.delete([file1, file2, file3])
+    #     Shrine.new(:storage).delete([file1, file2, file3])
     #
     # Now if you're using Storage::S3, deleting an array of files will issue a
     # single HTTP request. Some other storages may support multi deletes as
@@ -15,15 +15,6 @@ class Shrine
     # at once.
     module MultiDelete
       module InstanceMethods
-        # This allows `Shrine.delete` to accept an array of files.
-        def uploaded?(uploaded_file)
-          if uploaded_file.is_a?(Array)
-            uploaded_file.all? { |file| super(file) }
-          else
-            super
-          end
-        end
-
         private
 
         # Adds the ability to upload multiple files, leveraging the underlying