+++ lib/shrine/plugins/restore_cached.rb
@@ -21,8 +21,7 @@ class Shrine
           uploaded_file = uploaded_file(value) do |file|
             next unless cache.uploaded?(file)
             return unless file.exists?
-            uploader = shrine_class.uploader_for(file)
-            real_metadata = uploader.extract_metadata(file.to_io, context)
+            real_metadata = cache.extract_metadata(file.to_io, context)
             file.metadata.update(real_metadata)
           end