+++ lib/shrine/plugins/versions.rb
@@ -65,10 +65,10 @@ class Shrine
     # When deleting versions, any multi delete capabilities will be leveraged,
     # so when usingStorage::S3, deleting versions will issue only a single HTTP
     # request. If you want to delete versions manually, you can use
-    # `Shrine.delete`:
+    # `Shrine#delete`:
     #
     #     versions.keys #=> [:small, :medium, :large]
-    #     ImageUploader.delete(versions) # deletes a hash of versions
+    #     ImageUploader.new(:storage).delete(versions) # deletes a hash of versions
     module Versions
       def self.load_dependencies(uploader, *)
         uploader.plugin :multi_delete